+++ README.md
@@ -1,11 +1,73 @@
-# JSX
-1. Необходимо создать верстку для следующего списка:
-   1. Посмотреть вокруг
-   2. Оглянуться назад
-   3. Рыбов купить
-   4. "Красивое" сказать
+# tensor-school-frontend
+## Как писать тесты
+1. Создайте папку "lesson-x-y", где x - название раздела, y - название урока.
+1. Внутри папки создайте файл README.md с описанием задания.
+1. Создайте ts (или tsx, если нужно) файл с кодом, который вы ожидаете от студентов. Код должен быть полностью рабочим.
+    1. Если вы пишете на React и вам нужно посмотреть что строится, нужно временно заимпортить свой компонент в index.tsx и отрисовать его там. Студентам этого делать не нужно будет, пока мы просто собираем всё в одном месте.
+1. Рядом создайте файл test.ts (или tsx, если нужно) с таким же названием. В нём должны находиться тесты.
 
-С помощью уже созданных кнопок сделать смену:
-2. Типа списка на маркированный\нумерованный. Тип по умолчанию маркированный.
-3. Цвета текста на красный(red)\фиолетовый(violet). Цвет по умолчанию красный.
-4. Размера шрифта на большой\маленький (с помощью css). Шрифт по умолчанию большой.
+## Формат тестов
+1. Все названия должны быть максимально понятны студентам в случае падения тестов. Т.е. русский язык везде, где это имеет смысл, и названия должны отражать какое поведение проверяет тест.
+1. В корне describe должно быть название темы, например, JSX.
+```typescript
+describe('JSX', () => {
+    
+});
+```
+1. Если требуется разбивать на подтемы, нужно это делать через вложенные describe.
+```typescript
+describe('JSX', () => {
+    describe('Переключение стиля списка', () => {
+    });
+    describe('Переключение размера шрифта', () => {
+    });
+});
+```
+1. Тесты пишутся в блоках it. Один сценарий - один тест, не нужно закидывать все тесты в один it.
+```typescript
+describe('JSX', () => {
+    describe('Переключение стиля списка', () => {
+        it('Неупорядоченный список', () => {
+            const {container} = render(<App/>);
+
+            expect(container).toMatchSnapshot();
+        });
+
+        it('Упорядоченный список', () => {
+            const {queryByTestId, container} = render(<App/>);
+
+            const listTypeToggle = queryByTestId('listType') as HTMLElement;
+            fireEvent.click(listTypeToggle);
+
+            expect(container).toMatchSnapshot();
+        });
+    });
+});
+```
+
+## Советы по написанию тестов
+У нас используется фреймворк Jest. Большинству будет достаточно этого раздела документации: https://jestjs.io/docs/expect
+
+Также есть react-testing-library: https://testing-library.com/docs/react-testing-library/intro, если тестируете реактовские компоненты, то лучше использовать render и fireEvent оттуда.
+
+Теперь сами советы:
+1. Используйте максимально подходящие по смыслу проверки, не нужно всё проверять через toEqual, если можно написать более узкую проверку.
+    1. Если проверяете вёрстку, то в 99% случаев вам нужно `expect(dom).toMatchSnapshot()`;
+1. Пишите максимально изолированные тесты. Если вы меняете что-то глобальное - чистите за собой в блоке afterEach. Например, если задача требует руками редактировать DOM - точно нужно чистить.
+    1. Все вещи, отрендеренные через render из react-testing-library, автоматически удаляются после окончания теста.
+1. Если нужно отследить вызвал ли студент какой-то метод - используйте моки: https://jestjs.io/docs/mock-function-api
+```typescript
+jest.spyOn(window, 'alert');
+alert(123);
+expect(window.alert).toHaveBeenCalledWith(456);
+```
+Моки глобальные, не забываем чистить за собой:
+```typescript
+describe('...', () => {
+    afterEach(() => {
+        jest.restoreAllMocks();
+    });
+
+    it('...', () => {});
+});
+```
