+++ src/index.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
 import './index.css'
 
 ReactDOM.render(
     <React.StrictMode>
-      <App />
+        <div>123</div>
     </React.StrictMode>,
     document.getElementById('root')
 );
