+++ src/lesson-11-2/README.md
@@ -1,0 +1,11 @@
+# JSX
+1. Необходимо создать верстку для следующего списка:
+   1. Посмотреть вокруг
+   2. Оглянуться назад
+   3. Рыбов купить
+   4. "Красивое" сказать
+
+С помощью уже созданных кнопок сделать смену:
+2. Типа списка на маркированный\нумерованный. Тип по умолчанию маркированный.
+3. Цвета текста на красный(red)\фиолетовый(violet). Цвет по умолчанию красный.
+4. Размера шрифта на большой\маленький (с помощью css). Шрифт по умолчанию большой.
